requirements.txt
flask=3.1.0
flask-Cors=5.0.0
groq==0.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytest==7.4.3
black==23.12.1
isort==5.13.2
jupyter==1.0.0
pre-commit==3.5.0
pytest-cov==4.1.0
mypy==1.7.1